Add initial AssertJ assertions
As noted in #181, we can provide custom AssertJ assertions as a way to
simplify the ability to write unit tests against this library.

We use an `abstract` parent class to de-duplicate a number of common
calls across assertion classes.

We want to allow future improvements, which would assert based on the
level a log is, such as:

	assertThat(testLogger).hasLevel(ERROR).withMessage(...)

But for now, we can provide the "bare minimum" against what I've used
previously.

Closes #181.