Apply Spotless for consistent code formatting
To make it easier, and more consistent, for new contributors to the
project to write code that doesn't require style-related comments, we
can use the Spotless' Maven plugin to auto-format our code.

This uses the Google Java Code Style, overriding the default 2 spaces
with 4, as per project preferences, as well as adding our XML files to
be consistently formatted, too.

Additionally, add an inital `CONTRIBUTING.md` which details how to use
Spotless.

Closes #185.