Use abstract test classes in `TestLoggerAssertionsTest`
As part of a future change, we want to be able to write tests for both
the current method signatures we have, as well as against a
`LoggingEvent`.

To make it possible to reuse code for both situations, we can utilise an
abstract test class, which allows child test classes to implement the
specifics of their implementation (i.e. calling the right method on the
`assertions`), while taking advantage of the shared test setup +
verification.