Add assertions based on `LoggingEvent`
Asserting based on a `LoggingEvent` is actually a simpler API, and so
we should make it possible to use this, as well as the existing methods
we have available.

This requires we provide a helper to map the `Level` to a LoggingEvent`
with its arguments.

We can replace `TestLoggerAssert`'s methods with the new
`LoggingEvent`-based methods, as they are much clearer for their intent.

This also tweaks the error messages for failed assertions to utilise
`LoggingEvent.toString()`, as it provides a readable, descriptive
explanation of what was/wasn't available.