Introduce new setting to configure global capture level (#324)
Introduce a new global setting `capture.level` to disable storing (and
printing) logs at a given level (following the level hierarchy).

The implementation is very similar to the `print.level` global setting.

This is useful when tests are generating a lot of logging events that
are not of interest and don't need to be captured. ArchUnit tests are
known to be such tests and can be quite slow when slf4j-test is used.